refactor: rename files from .tsx to .ts and axios-api to axios-sdk
